Port dibbler to Android

+++ poslib/dnssec-sign.cpp
@@ -23,7 +23,7 @@
 #include "dnssec-sign.h"
 
 extern "C" {
-#include "hmac.h"
+#include "nettle/hmac.h"
 #include "base64.h"
 }
 
@@ -65,8 +65,8 @@ void verify_signature (DnsRR *check_tsig, DnsRR *message_tsig,
   u_int48 servertime = uint48_value (rr_getdata (check_tsig->RDATA, DNS_TYPE_TSIG, 1));
   uint16_t fudge = rr_getshort (check_tsig->RDATA, DNS_TYPE_TSIG, 2);
   
-  if ( (clienttime > servertime && clienttime - servertime > fudge) ||
-       (servertime > clienttime && servertime - clienttime > fudge) ) {
+  if ( (clienttime > servertime && (uint16_t)(clienttime - servertime) > fudge) ||
+       (servertime > clienttime && (uint16_t)(servertime - clienttime) > fudge) ) {
       throw PException (true, "Answer sign time invalid (answer time %d, real time %d)",
                         uint48_value (rr_getdata (message_tsig->RDATA, DNS_TYPE_TSIG, 1)),
                         clienttime);
